% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintchip.R
\docType{data}
\name{features_demo}
\alias{features_demo}
\title{Sample features for demo}
\format{
## `features_demo`
A data.table with 4 rows and 2 columns:
\describe{
  \item{start,end}{start and end of the sample features}
}
}
\source{
Xue et al.
}
\usage{
features_demo
}
\description{
Sample features for demo, shares some start and end locations with interactions
}
\keyword{datasets}
