% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintchip.R
\docType{data}
\name{frags_demo}
\alias{frags_demo}
\title{Sample fragments for demo}
\format{
## `frags_demo`
A data.table with 143 rows and 4 columns:
\describe{
  \item{start,end}{start and end of the sample interactions}
  \item{thickness}{example strand thicknesses for demo interactions}
  \item{annotation}{example annotations for demo interactions}
}
}
\source{
Xue et al.
}
\usage{
frags_demo
}
\description{
Sample interactions obtained from Xue et al.
}
\keyword{datasets}
