% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintchip.R
\docType{data}
\name{hg38_gene_locations}
\alias{hg38_gene_locations}
\title{Genomic build hg38 gencode gene locations}
\format{
## `hg38_gene_locations`
A data.table with 46,096 rows and 12 columns:
\describe{
  \item{seqnames}{Chromosome location}
  \item{start,end}{start and end of the gene from build hg19, includes UTRs}
  \item{strand}{Strand of the transcribed gene}
  \item{width}{Width of the transcribed gene}
  \item{gene_name}{Gene name}
  \item{transcript_name}{Transcript name}
  \item{transcript_id}{Gencode transcript ID}
  \item{type}{Artifact of the original dataset used to generate this file}
  \item{exon_number}{Artifact of the original dataset used to generate this file}
  \item{type.1}{Artifact of the original dataset used to generate this file}
  \item{tx.ord}{Artifact of the original dataset used to generate this file}
}
}
\source{
<http://mskilab.com/gTrack/hg38//gencode.composite.collapsed.rds>
}
\usage{
hg38_gene_locations
}
\description{
Gencode gene locations. Includes the exons, introns and UTRs
}
\details{
Code used to generate dataset:
install_github(mskilab/gTrack)
library(gTrack)
gt38 = track.gencode(build = 'hg38')
dat = gt38@data
dt_38 = mclapply(1:length(dat[[1]]), function(x){
  g = dat[[1]][[x]]
  gr2dt(g[g$type == 'gene'])
}, mc.cores = 8)
hg38_gene_locations = rbindlist(dt_38)
usethis::use_data(hg38_gene_locations)
}
\keyword{datasets}
