% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mintchip.R
\name{mintchip}
\alias{mintchip}
\title{mintchip, the Genomic-Inteaction Visualizer}
\usage{
mintchip(
  interactions = NULL,
  features = NULL,
  gene_list = NULL,
  genome_build = "hg38",
  alpha = 1,
  thickness_scale = 1,
  base_thickness = 1,
  xmin = NULL,
  xmax = NULL,
  color_palette = "Dark2",
  curvature = -0.25,
  number_of_points_for_drawing_line = 20
)
}
\arguments{
\item{interactions}{a data.frame or data.table that contains start
(default column name - 'start' and end (default column name = 'end') coordinates of
genomic interactions as well as any annotations you want to color by
(default column name - 'annotation') and a numeric column that will
map to interaction thickness (default column name - "thickness")
and a character column "strand" that can only contain ("+") or ("-") and that
indicates whether to plot the loops on top ("+") or on bottom ("-") note that this
stacks with curvature, so a positive curvature will invert the top and bottom.}

\item{features}{a data.frame or data.table that contains the
start (default column name - 'start') and end (default column name = 'end') coordinates of
genomic features that you would like to plot as well as the color you would like these
elements to be (default column name - 'color') and any text annotations that you
would like to add to the features (default column name - 'name')}

\item{gene_list}{a chacter vector of gencode gene names that you would like to plot
below the interactions, and above the genomic features}

\item{genome_build}{either "hg19" or "hg38" and indicates which genome build
to pull the gene annotations from}

\item{alpha}{a numeric value between 0 and 1 that specifies the transparency of the
genomic interactions. alpha scales linearly with the number of overlaps, e.g.
if alpha = 0.5 and two interactions intersect, their intersection will have the
same transparency as one interaction at alpha = 1.
alpha = 1 is no transparency
alpha = 0 is fully transparency}

\item{thickness_scale}{a numeric (0,Inf) that sets the amount by which the thicker peaks will stand out from thinner peaks
a high thickness_scale (e.g. 5) will yield thick lines that are much thicker than the thinner lines
a low thickness_scale (e.g. 0.1) will yield a plot where most of the lines are similar thickness}

\item{base_thickness}{linearly scales all line widths, e.g. if set to 2, all lines will be twice as thick
if set to 0.5, all lines will be half as thick}

\item{xmin}{a numeric that indicates the leftmost genomic coordinate for plotting viewing.
If not specified, automatically generated from the interactions, features, and gene_list.}

\item{xmax}{a numeric that indicates the rightmost genomic coordinate for plotting viewing.
If not specified, automatically generated from the interactions, features, and gene_list.}

\item{color_palette}{The RColorBrewer palette to use for annotating the
genomic interations, default is "Dark2"}

\item{curvature}{a numeric (-Inf,Inf) A numeric value giving the amount of curvature. 
Negative values produce left-hand curves, 
positive values produce right-hand curves, 
and zero produces a straight line. Taken from geom_curves documentation.}
}
\value{
a ggplot object that visualizes the provided genomic interactions and features
}
\description{
mintchip takes in a series of paired genomic interactions and features
and plots them within a given genomic window using ggplot2.
}
\author{
Zoran Z. Gajic
}
